Creating the custom useForm hook for the Form
